+++ js/data.js
@@ -1,0 +1,14 @@
+//тип объекта
+const TYPES_OF_PROPERTY = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+//время заезда
+const TIMES_OF_CHECKIN = ['12:00', '13:00', '14:00'];
+//время выезда
+const TIMES_OF_CHECKOUT = ['12:00', '13:00', '14:00'];
+//список возможных удобств
+const LIST_OF_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+//список фото обьектов
+const LIST_OF_PHOTOS = [
+  'https: //assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https: //assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https: //assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
